Introduce a ListMixin as an abstract base class for List nodes.
