+++ model/index.js
@@ -49,6 +49,7 @@ export { default as HTMLImporter } from './HTMLImporter'
 export { default as IncrementalData } from './IncrementalData'
 export { default as InlineNode } from './InlineNode'
 export { default as JSONConverter } from './JSONConverter'
+export { default as ListMixin } from './ListMixin'
 export { default as Marker } from './Marker'
 export { default as Node } from './Node'
 export { default as NodeIndex } from './NodeIndex'
